requirements.txt
click==8.1.2
Flask==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
torch==1.11.0
torchvision==0.10.0
typing_extensions==4.2.0
Werkzeug==2.1.1
